Version 0.27 - see releasenotes.txt for details!
V 0.27

- Selection of From/To systems for route planning via a drop-down-box
and changed message accordingly

- Different handling of routing cache, it will always load a cache file
on startup if it finds one. This will make startup about
one second longer but speeds up route finding substantially (~3-4x as
fast( on longer routes. You need the
neighbor-cache-28.0LY-dump file for this to work. If it's not present,
the app will still run as before.

- Routing in "save fuel" mode will announce doing so

- Advise the pilot accordingly when finding a route fails

- "Alternate Jumps" command will now look for jumps of at most 14 LY to
work around the "incomplete jump list" bug in E

- Ability to set the window to a standard, resizable format via an
(undocumented for now) .ini-setting. This won't be standard
for now since it leads to thicker window borders which I find to hurt
immersion.
#### Ini setting is: Resizable_window=True ###

- Brighter blue text in the messages (lower) window for better
readability